+++ components/net/lwip/Kconfig
@@ -1,4 +1,4 @@
-menuconfig RT_USING_LWIP 
+menuconfig RT_USING_LWIP
     bool "Enable light weight TCP/IP stack (lwIP)"
     select RT_USING_DEVICE
     default n
