+++ components/net/lwip/lwip-1.4.1/SConscript
@@ -10,7 +10,6 @@ src/api/netdb.c
 src/api/netifapi.c
 src/api/sockets.c
 src/api/tcpip.c
-src/arch/sys_arch.c
 src/core/def.c
 src/core/dhcp.c
 src/core/dns.c
@@ -68,7 +67,6 @@ src/netif/ppp/vj.c
 path = [GetCurrentDir() + '/src',
     GetCurrentDir() + '/src/include',
     GetCurrentDir() + '/src/include/ipv4',
-    GetCurrentDir() + '/src/arch/include',
     GetCurrentDir() + '/src/include/netif']
 
 if GetDepend(['RT_LWIP_SNMP']):
