+++ components/net/lwip/lwip-2.1.2/src/netif/ethernetif.c
@@ -42,6 +42,8 @@
  * 2021-09-07     Grissiom     fix eth_tx_msg ack bug
  */
 
+#include <string.h>
+
 #include "lwip/opt.h"
 #include "lwip/debug.h"
 #include "lwip/def.h"
@@ -53,11 +55,9 @@
 #include "lwip/tcpip.h"
 #include "lwip/dhcp.h"
 #include "lwip/netifapi.h"
-
+#include "lwip/inet.h"
 #include "netif/etharp.h"
 #include "netif/ethernetif.h"
-
-#include "lwip/inet.h"
 
 #include <ipc/completion.h>
