+++ components/net/lwip/port/arch/include/arch/bpstruct.h
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2006-2022, RT-Thread Development Team
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Change Logs:
+ * Date           Author       Notes
+ * 2022-02-20     Meco Man     add RT-Thread copyright
+ */
+
+#if defined(__ICCARM__)
+#pragma pack(1)
+#endif
